Added graphical method implementation for parametric images
A new function `generate_parametric_images_with_graphical_method` is introduced that allows the use of different graphical analysis methods when generating parametric images. The chosen method ("patlak", "logan", or "alt-logan") is passed as a string and the corresponding function is selected for analysis. This makes it easier to switch between different analysis methods.