Add smart_logan_analysis method and revise notebook tests
Added a `smart_logan_analysis` method that prevents division by 0 errors by only performing the calculations on points where the denominator is non-zero. We also now check to have at least 3 points being fit; if not, we return np.nan. The notebook was used to test the function (and the alt-logan function).