Refactor graphical_analysis.py for readability
Empty lines were added in between function definitions to improve readability of the code in the 'graphical_analysis.py' file. A redundant blank line in the GraphicalAnalysis class was also removed. Readability is key for code maintenance and these slight adjustments enhance the structure and visual clarity of the script.