Refactor pet_cli graphical plots for greater control and clarity
An updated version of the graphical_plots module in the pet_cli application has been introduced. In response to user feedback, a new function, 'generate_figure', has been added. This allows users to customize the graphical display of their data more easily and precisely. Additionally, an abstract method 'add_figure_labels' has been created to impose a uniform way of adding labels to the graphics. These changes are designed to make plot customization more accessible and straightforward for users, in addition to enhancing the visual clarity of the output.