Add Logan and AltLogan classes and update PatlakPlot
Implemented LoganPlot and AltLoganPlot as subclasses of the existing GraphicalAnalysisPlot class, adding corresponding methods to calculate values and set figure labels and legends. Refactored the calculate_x_and_y method in the PatlakPlot class. Also, added titles to linear and log-log plots and updated the yscale and xscale for the log-log plot. We now explicitly calculate the fit in the `calculate_x_and_y` method, which could be used to store the t_thresh index and non-zero indicies to take the redundant calculations out.