Add filepath and directory validation to Plot class
Introduced specific methods to validate filepaths and directories provided as arguments to the 'Plot' class in graphical_plots.py. Methods '_validate_filepath' and '_validate_directory' raise a ValueError if the provided string does not correspond to a valid file or directory. These validation steps ensure that the paths for TAC input, ROI and output directory are valid before plot generation.