Add function to safely write TAC file
The implemented '_safe_write_tac' function ensures a secure way of writing Time-Activity Curves (TACs) to files. With this function, the interpolated TAC can now be saved directly in the desired location specified by the user. Also, the '--outfile' argument has been renamed to '--out-tac-path' for clarity and the commented code for saving the file has been removed.