Add documentation for 'pet_cli.image_visualization' module
A file level documentation for 'pet_cli.image_visualization' is added in a new 'docs/pet_cli.image_visualization.rst' file. The documentation includes an overview of the module and the classes defined within. The 'pet_cli.image_visualization' module is also added to the 'docs/modules.rst' file. Further, the docstrings in the 'pet_cli.image_visualization' module are updated to improve clarity and readability.