Update NiftiGifCreator description in image_visualization
The code comments for NiftiGifCreator in image_visualization.py have been updated. The class is now noted to support 'transverse' view when creating a GIF from a NIfTI image. Additionally, it is now emphasized that the class makes a GIF of the image in the space defined by the nifty file.