Add option for specifying time frame keyword in image_operations_4d.py
An option is added to specify the time frame keyword (either 'FrameReferenceTime' or 'FrameTimesStart') in the function 'write_TACs' in the image_operations_4d.py file. The code also checks if the provided time frame keyword is valid. Additionally, some refactoring was done to improve code clarity, such as shifting the writing of JSON output to the 'write_dict_to_json' method within the ImageIO class.