Refactor and correct calculations in reference_tissue_models.py
Performed a set of corrections and refactors in the reference tissue models Python file. Fixed bugs in several function computations and improved the clarity of the code. Also, the way fit parameters are formatted and returned in the _calc_frtm_or_srtm_fit_props method was modified for better readability.