Added a safe TAC loading function (which should be replaced in the future).
This update introduces a new function, `_safe_load_tac`, which is responsible for loading time-activity curves (TAC) from files. This function takes a file name as input, attempts to load a TAC from the specified file, and raises an exception if it encounters an error. The returned result is a numpy array with two indexes representing times and activity respectively.