Rearrange argument declarations and add weighting decay constant parameter
Rearranged arguments declaration under 'Analysis Parameters' in the correct alphabetical order for better readability. Also introduced a new parameter, 'weighting decay constant' for computing per-frame weighting for fits. This new parameter is not required and will be used to further customize the fitting process. Note that since TACFitter does not implement reading in a weights file, neither does the CLI for now.