Add pretty print functionality to CLI for fit results
A new "_pretty_print_results" function was introduced to format and display the output from parameter fit procedures. This function is called within the main function if the "print" argument is passed. It shows parameter names along with their fit values, standard errors, and percent errors. Additional logic was also added to handle the printing of Binding Potential (BP) and k2_prime specifically for MRTM methods.